TFP-6240: Trekker restkall ut i egen task
